+++ src/main/java/web/controller/Controller.java
@@ -1,7 +1,7 @@
 package web.controller;
 
-import web.HttpRequest;
-import web.HttpResponse;
+import web.request.HttpRequest;
+import web.response.HttpResponse;
 
 public interface Controller {
     void doService(HttpRequest httpRequest, HttpResponse httpResponse);
