+++ src/main/java/web/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package web;
+package web.request;
 
 public enum HttpMethod {
     GET,
