+++ src/main/java/web/request/RequestUri.java
@@ -1,4 +1,4 @@
-package web;
+package web.request;
 
 public class RequestUri {
     private static final int METHOD_INDEX = 0;
