+++ src/main/java/web/response/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package web;
+package web.response;
 
 public enum HttpStatusCode {
     OK("200 OK"),
