+++ src/test/java/web/RequestPathTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import web.request.RequestPath;
 
 class RequestPathTest {
